REL-2859 Properly extract the wavelength for GNIRS and AcqCam and some refactoring #time 2h
